refactor: askCarNames 함수 위치 이동
