+++ src/controller.js
@@ -1,16 +1,5 @@
-import { readLineAsync } from "./utils/readLineSync.js";
 import Race from "./domain/Race.js";
 import { displayWinners, displayRace } from "./view.js";
-
-export async function askCarNames() {
-  const carNames = await readLineAsync(
-    "경주할 자동차 이름을 입력하세요. (이름은 쉼표(,) 기준으로 구분)"
-  );
-
-  console.log(carNames);
-  console.log("");
-  return carNames;
-}
 
 export function playGame(carNames) {
   try {
