+++ src/index.js
@@ -1,7 +1,9 @@
-import { askCarNames, playGame } from "./controller.js";
+import { askCarNames } from "./view.js";
+import { playGame } from "./controller.js";
 
 async function app() {
   const carNames = await askCarNames();
+
   playGame(carNames);
 }
